requirements.txt
jupyterlab
pandas
sklearn
flask
request
